Fix raw usage of server
Since we can't know what kind of context and request
on each of all the servers being used in the setup, we
now return an Object instead of the Server interface.

This does not break anything since the end-user already today
needs to cast the returned server to the correct protocol server
interface.